tests: merged test
